// src/app/pages/chofers/components/chofers-table/chofers.interface.ts
export interface ChofersInterface {
   idchofer?: number;
   licencia?: string;
   fianza?: number;
   status?: string;
   chofer?: number;
   aval1?: number;
   aval2?: number;
   aval3?: number;
   aval4?: number;
   baja?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}